requirements.txt
streamlit
pandas
numpy
matplotlib
openpyxl
python-dotenv
